feat(auth/limiter): split limiters on reset password
- resetPassword request will not be blocked, only limited
- update password is secured against brute force attack with blocking
- only failed attempts count toward update password limiter

+++ app/controllers/users_controller.ts
@@ -2,6 +2,7 @@ import { inject } from "@adonisjs/core";
 import type { HttpContext } from "@adonisjs/core/http";
 
 import ResetPasswordService from "#services/reset_password_service";
+import { updatePasswordLimiter } from "#start/limiter";
 import { emailValidator } from "#validators/email";
 import { passwordValidator } from "#validators/password";
 import { resetPasswordTokenValidator } from "#validators/reset_password_token";
@@ -31,11 +32,25 @@ export default class UsersController {
     { request, response }: HttpContext,
     resetPasswordService: ResetPasswordService,
   ) {
-    await request.validateUsing(resetPasswordTokenValidator);
+    const { limiter, errorMessage } = updatePasswordLimiter;
+    const key = `update_password_${request.ip()}`;
+
+    // Error raised on limiter exhaust. Only failed attempts count.
+    // Fail on `.validateUsing()` will proceed with response immediately
+    const [error, result] = await limiter.penalize(key, async () => {
+      return await request.validateUsing(resetPasswordTokenValidator);
+    });
+
+    if (error !== null) {
+      return response.tooManyRequests({
+        message: errorMessage,
+        retryAfter: error.response.availableIn,
+      });
+    }
 
     const {
       params: { token },
-    } = await request.validateUsing(resetPasswordTokenValidator);
+    } = result;
 
     if (!token.isValid()) {
       await resetPasswordService.destroyToken(token.user);
